acsess/acsess_backend/requirements.txt
django>=3.2,<5.0.6
djangorestframework>=3.12.4,<4.0.0
django-cors-headers>=3.8.0,<5.0.0
psycopg2-binary>=2.8.6,<3.0.0
pandas
openpyxl
serializer
djangorestframework
djangorestframework-simplejwt
celery>=5.1,<6.0
redis>=4.0,<5.0
pytest-django
drf-yasg
django-celery-beat
django-extensions
graphviz
python-decouple